🔧 Change fterm border with single from double (config/nvim/lua/plugin/settings/fterm.lua)

+++ config/nvim/lua/plugin/settings/fterm.lua
@@ -1,8 +1,7 @@
 require("FTerm").setup({
-	border = "double",
+	border = "single",
 	dimensions = {
 		height = 0.9,
 		width = 0.9,
 	},
 })
-
